consolidate ui modules
